Remove useless EqualsAndHashCode in ColumnMetaData, ShardingSphereColumn and other objects
