+++ infra/database/core/src/main/java/org/apache/shardingsphere/infra/database/core/metadata/data/model/SchemaMetaData.java
@@ -17,7 +17,6 @@
 
 package org.apache.shardingsphere.infra.database.core.metadata.data.model;
 
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
@@ -29,7 +28,6 @@ import java.util.Collection;
 */
 @RequiredArgsConstructor
 @Getter
-@EqualsAndHashCode
 @ToString
 public final class SchemaMetaData {
